Fix TodoService companion object shadowing
